Switch to pywin32 UI
